+++ index.php
@@ -1,6 +1,7 @@
 <?php
 date_default_timezone_set('Europe/Moscow');
-require_once('functions.php');
+require_once('functions/db.php');
+require_once('functions/template.php');
 $config = require('config.php');
 
 $is_auth = rand(0, 1);
